+++ cypress-shared/support/common/constants.js
@@ -43,7 +43,7 @@ export const FAIL_ON_STATUS_CODE = {
   failOnStatusCode: false,
 }
 
-export const FAIL_ON_STATUS_CODE_STRING = 'failOnStatusCode: false'
+export const FAIL_ON_STATUS_CODE_STRING = 'failOnStatusCode: false,'
 
 export const ENTITIES = {
   CLIENTS: { id: 'CL', searchKey: 'email' },