+++ cypress-shared/support/common/support.js
@@ -498,11 +498,11 @@ function logic(storeFrontCookie, stop) {
       }
     })
   })
-  beforeEach(() => {
+  beforeEach(function () {
     cy.qe(
       `Testcase title - ${
         Cypress.mocha.getRunner().suite.ctx.currentTest.title
-      }`
+      }, Attempt Number - ${this.currentTest.currentRetry() + 1}`
     )
   })
   if (stop) stopTestCaseOnFailure()